# artifacts/api-server/src/routes/admin.ts
import { Router } from "express";
import { requireAuth } from "../middlewares/auth";
import {
  isCardProgramEnabled, setCardProgramEnabled,
  getMaintenance, setMaintenance,
  getFeeSchedule, setFeeSchedule, type FeeSchedule,
  getWalletProviderConfig, setWalletProviderConfig,
  WALLET_PROVIDER_KEYS, type WalletProviderKey,
} from "../lib/settings";
import { walletProviderCatalog } from "../lib/wallet-providers";
import { isStripeConfigured } from "../lib/stripe-issuing";
import { invalidateCustomJobs, CATEGORY_LABELS } from "../lib/jobs";
import { db, usersTable, transactionsTable, cardWaitlistTable, customJobsTable } from "@workspace/db";
import { eq, count, desc, sql, isNotNull } from "drizzle-orm";

const router = Router();

// Comma-separated admin emails configured via env var
const ADMIN_EMAILS = (process.env.ADMIN_EMAILS ?? "admin@spayewallet.com")
  .split(",").map((e) => e.trim().toLowerCase());

function requireAdmin(req: any, res: any, next: any) {
  if (!req.user) { res.status(401).json({ error: "unauthorized" }); return; }
  if (!ADMIN_EMAILS.includes(req.user.email.toLowerCase())) {
    res.status(403).json({ error: "forbidden", message: "Admin access required" });
    return;
  }
  next();
}

router.get("/admin/stats", requireAuth, requireAdmin, async (req, res) => {
  try {
    const [{ total: totalUsers }] = await db.select({ total: count() }).from(usersTable);
    const [{ total: pendingKyc }] = await db.select({ total: count() }).from(usersTable).where(eq(usersTable.kycStatus, "pending"));
    const [{ total: approvedKyc }] = await db.select({ total: count() }).from(usersTable).where(eq(usersTable.kycStatus, "approved"));
    const [{ total: rejectedKyc }] = await db.select({ total: count() }).from(usersTable).where(eq(usersTable.kycStatus, "rejected"));

    const today = new Date();
    today.setHours(0, 0, 0, 0);
    const [{ total: transactionsToday }] = await db.select({ total: count() }).from(transactionsTable)
      .where(sql`${transactionsTable.createdAt} >= ${today}`);

    const [volumeRow] = await db.select({
      volume: sql<string>`COALESCE(SUM(${transactionsTable.amount}::numeric), 0)`,
    }).from(transactionsTable);

    const thirtyDaysAgo = new Date(Date.now() - 30 * 24 * 60 * 60 * 1000);
    const [{ total: activeUsers }] = await db.select({ total: count() }).from(usersTable)
      .where(sql`${usersTable.createdAt} >= ${thirtyDaysAgo}`);

    const [{ total: businessUsers }] = await db.select({ total: count() }).from(usersTable)
      .where(eq(usersTable.accountType, "business"));

    // Acquisition breakdown: where signups come from (jobs, landing, google, direct…)
    const sourceRows = await db.select({
      source: sql<string>`COALESCE(${usersTable.signupSource}, 'unknown')`,
      total: count(),
    }).from(usersTable).groupBy(sql`COALESCE(${usersTable.signupSource}, 'unknown')`);
    const signupsBySource: Record<string, number> = {};
    for (const row of sourceRows) {
      // Collapse "jobs:rv-123" style sources into their channel for the summary
      const channel = row.source.split(":")[0] || "unknown";
      signupsBySource[channel] = (signupsBySource[channel] ?? 0) + row.total;
    }

    res.json({
      totalUsers,
      activeUsers,
      businessUsers,
      pendingKyc,
      approvedKyc,
      rejectedKyc,
      transactionsToday,
      totalTransactionVolume: parseFloat(volumeRow?.volume ?? "0"),
      totalWalletBalance: 0,
      signupsBySource,
    });
  } catch (err) {
    req.log.error({ err }, "Admin stats error");
    res.status(500).json({ error: "internal_error", message: "Failed to fetch stats" });
  }
});

router.get("/admin/users", requireAuth, requireAdmin, async (req, res) => {
  try {
    const kycStatus = typeof req.query["kycStatus"] === "string" ? req.query["kycStatus"] : undefined;

    const baseQuery = db.select({
      id: usersTable.id,
      email: usersTable.email,
      fullName: usersTable.fullName,
      phoneNumber: usersTable.phoneNumber,
      country: usersTable.country,
      kycStatus: usersTable.kycStatus,
      avatarUrl: usersTable.avatarUrl,
      signupSource: usersTable.signupSource,
      accountType: usersTable.accountType,
      businessName: usersTable.businessName,
      createdAt: usersTable.createdAt,
    }).from(usersTable);

    const rows = await (
      kycStatus && ["pending", "approved", "rejected"].includes(kycStatus)
        ? baseQuery.where(eq(usersTable.kycStatus, kycStatus as "pending" | "approved" | "rejected"))
        : baseQuery
    ).orderBy(desc(usersTable.createdAt)).limit(200);

    const [{ total }] = await db.select({ total: count() }).from(usersTable);

    res.json({
      users: rows.map((u) => ({
        ...u,
        walletBalance: 0,
        createdAt: u.createdAt.toISOString(),
      })),
      total,
    });
  } catch (err) {
    req.log.error({ err }, "Admin users error");
    res.status(500).json({ error: "internal_error", message: "Failed to fetch users" });
  }
});

router.get("/admin/transactions", requireAuth, requireAdmin, async (req, res) => {
  try {
    const type = typeof req.query["type"] === "string" ? req.query["type"] : undefined;

    const rows = await (
      type && ["send", "receive", "withdraw", "recharge", "payment"].includes(type)
        ? db.select().from(transactionsTable).where(eq(transactionsTable.type, type as any))
        : db.select().from(transactionsTable)
    ).orderBy(desc(transactionsTable.createdAt)).limit(200);

    const [{ total }] = await db.select({ total: count() }).from(transactionsTable);

    res.json({
      transactions: rows.map((t) => ({
        ...t,
        amount: parseFloat(String(t.amount)),
        createdAt: t.createdAt.toISOString(),
      })),
      total,
    });
  } catch (err) {
    req.log.error({ err }, "Admin transactions error");
    res.status(500).json({ error: "internal_error", message: "Failed to fetch transactions" });
  }
});

// ─── Feature flags: the admin master switches ─────────────────────────────────

async function featureFlagsPayload() {
  const [cardProgramEnabled, maintenance, [{ total: cardWaitlistCount }]] = await Promise.all([
    isCardProgramEnabled(),
    getMaintenance(),
    db.select({ total: count() }).from(cardWaitlistTable),
  ]);
  return {
    cardProgramEnabled,
    stripeConfigured: isStripeConfigured(),
    cardWaitlistCount,
    maintenanceMode: maintenance.enabled,
    maintenanceMessage: maintenance.message,
  };
}

router.get("/admin/feature-flags", requireAuth, requireAdmin, async (req, res) => {
  try {
    res.json(await featureFlagsPayload());
  } catch (err) {
    req.log.error({ err }, "Feature flags read error");
    res.status(500).json({ error: "internal_error", message: "Failed to read feature flags" });
  }
});

router.put("/admin/feature-flags", requireAuth, requireAdmin, async (req, res) => {
  try {
    const { cardProgramEnabled, maintenanceMode, maintenanceMessage } = req.body as {
      cardProgramEnabled?: unknown; maintenanceMode?: unknown; maintenanceMessage?: unknown;
    };
    if (cardProgramEnabled === undefined && maintenanceMode === undefined && maintenanceMessage === undefined) {
      res.status(400).json({ error: "validation_error", message: "Provide at least one flag to update" });
      return;
    }
    if (cardProgramEnabled !== undefined) {
      if (typeof cardProgramEnabled !== "boolean") {
        res.status(400).json({ error: "validation_error", message: "cardProgramEnabled must be a boolean" });
        return;
      }
      await setCardProgramEnabled(cardProgramEnabled);
      req.log.info({ cardProgramEnabled, admin: req.user!.email }, "Card program switch changed");
    }
    if (maintenanceMode !== undefined || maintenanceMessage !== undefined) {
      if (maintenanceMode !== undefined && typeof maintenanceMode !== "boolean") {
        res.status(400).json({ error: "validation_error", message: "maintenanceMode must be a boolean" });
        return;
      }
      if (maintenanceMessage !== undefined && typeof maintenanceMessage !== "string") {
        res.status(400).json({ error: "validation_error", message: "maintenanceMessage must be a string" });
        return;
      }
      await setMaintenance({
        ...(maintenanceMode !== undefined ? { enabled: maintenanceMode } : {}),
        ...(maintenanceMessage !== undefined ? { message: maintenanceMessage.slice(0, 280) } : {}),
      });
      req.log.warn({ maintenanceMode, admin: req.user!.email }, "Maintenance mode changed");
    }
    res.json(await featureFlagsPayload());
  } catch (err) {
    req.log.error({ err }, "Feature flags update error");
    res.status(500).json({ error: "internal_error", message: "Failed to update feature flags" });
  }
});

// ─── Wallet provider switches (WaaS: Privy / Coinbase CDP / Turnkey) ──────────
// The active provider creates NEW wallets; per-provider toggles are kill
// switches (OFF = no new wallets and no sends signed via it). Existing wallets
// always keep the provider that holds their key — switching never moves funds.

async function walletProvidersPayload() {
  const config = await getWalletProviderConfig();

  // How many user wallets each provider holds keys for (legacy rows = Privy)
  const countRows = await db.select({
    provider: sql<string>`COALESCE(${usersTable.walletProvider}, 'privy')`,
    total: count(),
  }).from(usersTable)
    .where(isNotNull(usersTable.celoWalletAddress))
    .groupBy(sql`COALESCE(${usersTable.walletProvider}, 'privy')`);
  const walletCounts: Record<string, number> = {};
  for (const row of countRows) walletCounts[row.provider] = row.total;

  return {
    activeProvider: config.activeProvider,
    providers: walletProviderCatalog().map((p) => ({
      key: p.key,
      label: p.label,
      configured: p.configured,
      enabled: config.enabled[p.key],
      wallets: walletCounts[p.key] ?? 0,
      envHint: p.envHint,
    })),
  };
}

router.get("/admin/wallet-providers", requireAuth, requireAdmin, async (req, res) => {
  try {
    res.json(await walletProvidersPayload());
  } catch (err) {
    req.log.error({ err }, "Wallet providers read error");
    res.status(500).json({ error: "internal_error", message: "Failed to read wallet provider settings" });
  }
});

router.put("/admin/wallet-providers", requireAuth, requireAdmin, async (req, res) => {
  try {
    const { activeProvider, enabled } = req.body as { activeProvider?: unknown; enabled?: unknown };
    if (activeProvider === undefined && enabled === undefined) {
      res.status(400).json({ error: "validation_error", message: "Provide activeProvider and/or enabled to update" });
      return;
    }
    if (activeProvider !== undefined && !WALLET_PROVIDER_KEYS.includes(activeProvider as WalletProviderKey)) {
      res.status(400).json({ error: "validation_error", message: `activeProvider must be one of: ${WALLET_PROVIDER_KEYS.join(", ")}` });
      return;
    }
    const enabledUpdate: Partial<Record<WalletProviderKey, boolean>> = {};
    if (enabled !== undefined) {
      if (typeof enabled !== "object" || enabled === null) {
        res.status(400).json({ error: "validation_error", message: "enabled must be an object of provider→boolean" });
        return;
      }
      for (const [key, value] of Object.entries(enabled as Record<string, unknown>)) {
        if (!WALLET_PROVIDER_KEYS.includes(key as WalletProviderKey)) {
          res.status(400).json({ error: "validation_error", message: `Unknown provider "${key}"` });
          return;
        }
        if (typeof value !== "boolean") {
          res.status(400).json({ error: "validation_error", message: `enabled.${key} must be a boolean` });
          return;
        }
        enabledUpdate[key as WalletProviderKey] = value;
      }
    }
    await setWalletProviderConfig({
      ...(activeProvider !== undefined ? { activeProvider: activeProvider as WalletProviderKey } : {}),
      ...(enabled !== undefined ? { enabled: enabledUpdate } : {}),
    });
    req.log.warn({ activeProvider, enabled: enabledUpdate, admin: req.user!.email }, "Wallet provider switches changed");
    res.json(await walletProvidersPayload());
  } catch (err) {
    req.log.error({ err }, "Wallet providers update error");
    res.status(500).json({ error: "internal_error", message: "Failed to update wallet provider settings" });
  }
});

// ─── Fee schedule: provider cost + S-PAY margin = user price ──────────────────

router.get("/admin/fees", requireAuth, requireAdmin, async (req, res) => {
  try {
    res.json(await getFeeSchedule());
  } catch (err) {
    req.log.error({ err }, "Fee schedule read error");
    res.status(500).json({ error: "internal_error", message: "Failed to read fee schedule" });
  }
});

router.put("/admin/fees", requireAuth, requireAdmin, async (req, res) => {
  try {
    const body = req.body as Partial<FeeSchedule>;
    const fields: (keyof FeeSchedule)[] = ["withdrawalFeePercent", "withdrawalFeeMin", "cardIssuanceFee", "p2pFeePercent"];
    for (const f of fields) {
      const v = body[f];
      if (typeof v !== "number" || !Number.isFinite(v) || v < 0 || v > 100) {
        res.status(400).json({ error: "validation_error", message: `${f} must be a number between 0 and 100` });
        return;
      }
    }
    const fees: FeeSchedule = {
      withdrawalFeePercent: body.withdrawalFeePercent!,
      withdrawalFeeMin: body.withdrawalFeeMin!,
      cardIssuanceFee: body.cardIssuanceFee!,
      p2pFeePercent: body.p2pFeePercent!,
    };
    await setFeeSchedule(fees);
    req.log.info({ fees, admin: req.user!.email }, "Fee schedule updated");
    res.json(fees);
  } catch (err) {
    req.log.error({ err }, "Fee schedule update error");
    res.status(500).json({ error: "internal_error", message: "Failed to update fee schedule" });
  }
});

// ─── S-PAY's own job listings (pinned to the feed top; SEO content) ───────────

router.get("/admin/custom-jobs", requireAuth, requireAdmin, async (req, res) => {
  try {
    const rows = await db.select().from(customJobsTable)
      .where(eq(customJobsTable.active, true))
      .orderBy(desc(customJobsTable.createdAt));
    res.json({ jobs: rows.map((r) => ({ ...r, createdAt: r.createdAt.toISOString() })) });
  } catch (err) {
    req.log.error({ err }, "Custom jobs list error");
    res.status(500).json({ error: "internal_error", message: "Failed to list custom jobs" });
  }
});

router.post("/admin/custom-jobs", requireAuth, requireAdmin, async (req, res) => {
  try {
    const { title, company, applyUrl, category, location, salary, description } = req.body as Record<string, unknown>;
    if (typeof title !== "string" || !title.trim() || typeof company !== "string" || !company.trim()) {
      res.status(400).json({ error: "validation_error", message: "title and company are required" });
      return;
    }
    if (typeof applyUrl !== "string" || !/^https?:\/\//.test(applyUrl)) {
      res.status(400).json({ error: "validation_error", message: "applyUrl must be a valid http(s) URL" });
      return;
    }
    const cat = typeof category === "string" && (CATEGORY_LABELS as readonly string[]).includes(category)
      ? category : "Engineering";
    const [row] = await db.insert(customJobsTable).values({
      title: title.trim().slice(0, 160),
      company: company.trim().slice(0, 120),
      applyUrl: applyUrl.trim().slice(0, 500),
      category: cat,
      location: typeof location === "string" && location.trim() ? location.trim().slice(0, 120) : "Worldwide",
      salary: typeof salary === "string" && salary.trim() ? salary.trim().slice(0, 80) : "Competitive",
      description: typeof description === "string" && description.trim() ? description.trim().slice(0, 20000) : null,
    }).returning();
    invalidateCustomJobs();
    req.log.info({ jobId: row.id, admin: req.user!.email }, "Custom job listing created");
    res.json({ ...row, createdAt: row.createdAt.toISOString() });
  } catch (err) {
    req.log.error({ err }, "Custom job create error");
    res.status(500).json({ error: "internal_error", message: "Failed to create listing" });
  }
});

router.delete("/admin/custom-jobs/:jobId", requireAuth, requireAdmin, async (req, res) => {
  try {
    await db.update(customJobsTable)
      .set({ active: false })
      .where(eq(customJobsTable.id, req.params.jobId as string));
    invalidateCustomJobs();
    res.json({ message: "Listing removed from the feed." });
  } catch (err) {
    req.log.error({ err }, "Custom job delete error");
    res.status(500).json({ error: "internal_error", message: "Failed to remove listing" });
  }
});

// Returns which env vars are configured (not their values)
router.get("/admin/settings", requireAuth, requireAdmin, (req, res) => {
  const check = (key: string) => !!process.env[key];
  res.json({
    database: { configured: check("DATABASE_URL") },
    auth: {
      jwtConfigured: check("JWT_SECRET"),
      googleConfigured: check("GOOGLE_CLIENT_ID") && check("GOOGLE_CLIENT_SECRET"),
    },
    noah: {
      configured: check("NOAH_API_KEY"),
      webhookConfigured: check("NOAH_WEBHOOK_SECRET"),
    },
    wallet: Object.fromEntries(walletProviderCatalog().map((p) => [p.key, { configured: p.configured }])),
    stripe: {
      configured: check("STRIPE_SECRET_KEY"),
      webhookConfigured: check("STRIPE_WEBHOOK_SECRET"),
    },
    cors: {
      origin: process.env.CORS_ORIGIN ?? "(not set)",
    },
    adminEmails: ADMIN_EMAILS,
  });
});

export default router;