# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: S-PAY API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/stats:
    get:
      operationId: getAdminStats
      summary: Admin platform statistics
      tags: [Admin]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Admin stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
  /admin/users:
    get:
      operationId: getAdminUsers
      summary: List all users
      tags: [Admin]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: kycStatus
          schema:
            type: string
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUsersResponse"
  /admin/transactions:
    get:
      operationId: getAdminTransactions
      summary: List all transactions
      tags: [Admin]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: type
          schema:
            type: string
      responses:
        "200":
          description: Transactions list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminTransactionsResponse"
  /admin/settings:
    get:
      operationId: getAdminSettings
      summary: System configuration status
      tags: [Admin]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Settings status
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
  /admin/fees:
    get:
      operationId: getFeeSchedule
      summary: Read the platform fee schedule (provider cost + S-PAY margin = user price)
      tags: [Admin]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current fee schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeeSchedule"
    put:
      operationId: updateFeeSchedule
      summary: Update the platform fee schedule (applies to new quotes immediately, no deploy)
      tags: [Admin]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeeSchedule"
      responses:
        "200":
          description: Updated fee schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeeSchedule"
  /admin/custom-jobs:
    get:
      operationId: getCustomJobs
      summary: List S-PAY's own injected job listings
      tags: [Admin]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Custom listings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomJobsResponse"
    post:
      operationId: createCustomJob
      summary: Inject an S-PAY job listing (pinned to the top of the feed)
      tags: [Admin]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomJobCreateRequest"
      responses:
        "200":
          description: Created listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomJob"
  /admin/custom-jobs/{jobId}:
    delete:
      operationId: deleteCustomJob
      summary: Remove an injected listing
      tags: [Admin]
      security:
        - bearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/wallet-providers:
    get:
      operationId: getWalletProviders
      summary: Read the wallet (WaaS) provider switches — active provider, per-provider on/off, configuration status, wallet counts
      tags: [Admin]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current wallet provider configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletProvidersResponse"
    put:
      operationId: updateWalletProviders
      summary: Switch the active wallet provider and/or toggle providers on/off (live, no deploy)
      tags: [Admin]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WalletProvidersUpdateRequest"
      responses:
        "200":
          description: Updated wallet provider configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletProvidersResponse"
  /admin/feature-flags:
    get:
      operationId: getFeatureFlags
      summary: Read runtime feature flags (card program switch, readiness, waitlist demand)
      tags: [Admin]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current flags
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeatureFlagsResponse"
    put:
      operationId: updateFeatureFlags
      summary: Toggle runtime feature flags (e.g. switch the card program on)
      tags: [Admin]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeatureFlagsUpdateRequest"
      responses:
        "200":
          description: Updated flags
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeatureFlagsResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    FeeSchedule:
      type: object
      properties:
        withdrawalFeePercent:
          type: number
          description: User-facing withdrawal fee in percent (covers Noah's cost + S-PAY margin)
        withdrawalFeeMin:
          type: number
          description: Minimum withdrawal fee in USD
        cardIssuanceFee:
          type: number
          description: One-time virtual card creation fee in USD (covers Stripe's cost + margin)
        p2pFeePercent:
          type: number
          description: Internal S-PAY-to-S-PAY transfer fee in percent (0 = free, growth-friendly)
      required:
        - withdrawalFeePercent
        - withdrawalFeeMin
        - cardIssuanceFee
        - p2pFeePercent
    CustomJob:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        company:
          type: string
        category:
          type: string
        location:
          type: string
        salary:
          type: string
        applyUrl:
          type: string
        description:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - title
        - company
        - applyUrl
        - createdAt
    CustomJobsResponse:
      type: object
      properties:
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/CustomJob"
      required:
        - jobs
    CustomJobCreateRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 160
        company:
          type: string
          maxLength: 120
        applyUrl:
          type: string
          maxLength: 500
        category:
          type: string
          enum: [Engineering, Design, Marketing, Product, Sales, Finance, Operations]
        location:
          type: string
          maxLength: 120
        salary:
          type: string
          maxLength: 80
        description:
          type: string
          maxLength: 20000
      required:
        - title
        - company
        - applyUrl
    FeatureFlagsResponse:
      type: object
      properties:
        cardProgramEnabled:
          type: boolean
          description: Master switch for the virtual card program (admin-controlled)
        stripeConfigured:
          type: boolean
          description: Whether STRIPE_SECRET_KEY is set — cards can only be issued when true
        cardWaitlistCount:
          type: integer
          description: How many users are waiting for the card
        maintenanceMode:
          type: boolean
          description: When true, the API serves 503 for user traffic (admin + auth + webhooks stay up)
        maintenanceMessage:
          type: string
          description: Message shown to users while in maintenance
      required:
        - cardProgramEnabled
        - stripeConfigured
        - cardWaitlistCount
        - maintenanceMode
    FeatureFlagsUpdateRequest:
      type: object
      description: Partial update — provide any subset of flags
      properties:
        cardProgramEnabled:
          type: boolean
        maintenanceMode:
          type: boolean
        maintenanceMessage:
          type: string
          maxLength: 280
    WalletProviderInfo:
      type: object
      properties:
        key:
          type: string
          enum: [privy, cdp, turnkey]
        label:
          type: string
          description: Human-readable provider name
        configured:
          type: boolean
          description: Whether the provider's API credentials are set in the environment
        enabled:
          type: boolean
          description: Admin on/off switch — OFF blocks new wallets AND sends signed via this provider
        wallets:
          type: integer
          description: How many user wallets this provider currently holds keys for
        envHint:
          type: string
          description: The environment variables that activate this provider
      required:
        - key
        - label
        - configured
        - enabled
        - wallets
        - envHint
    WalletProvidersResponse:
      type: object
      properties:
        activeProvider:
          type: string
          enum: [privy, cdp, turnkey]
          description: Provider used to create NEW wallets (existing wallets always keep the provider that holds their key)
        providers:
          type: array
          items:
            $ref: "#/components/schemas/WalletProviderInfo"
      required:
        - activeProvider
        - providers
    WalletProvidersUpdateRequest:
      type: object
      description: Partial update — switch the active provider and/or toggle providers on/off
      properties:
        activeProvider:
          type: string
          enum: [privy, cdp, turnkey]
        enabled:
          type: object
          description: "Per-provider on/off, e.g. {\"privy\": false, \"cdp\": true}"
          properties:
            privy:
              type: boolean
            cdp:
              type: boolean
            turnkey:
              type: boolean
...
    AdminStats:
      type: object
      properties:
        totalUsers:
          type: integer
        activeUsers:
          type: integer
        totalTransactionVolume:
          type: number
        transactionsToday:
          type: integer
        pendingKyc:
          type: integer
        approvedKyc:
          type: integer
        rejectedKyc:
          type: integer
        totalWalletBalance:
          type: number
        signupsBySource:
          type: object
          description: "Signup counts grouped by acquisition source (jobs, landing, google, direct...)"
          additionalProperties:
            type: integer
        businessUsers:
          type: integer
          description: Number of business (KYB) accounts
      required:
        - totalUsers
        - activeUsers
        - totalTransactionVolume
        - transactionsToday
        - pendingKyc
        - approvedKyc
        - rejectedKyc
        - totalWalletBalance
    AdminUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        fullName:
          type: string
        kycStatus:
          type: string
        country:
          type: string
        walletBalance:
          type: number
        signupSource:
          type: string
          nullable: true
        accountType:
          type: string
          enum: [personal, business]
        businessName:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - email
        - fullName
        - kycStatus
    AdminUsersResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/AdminUser"
        total:
          type: integer
      required:
        - users
        - total
...
    AdminTransaction:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        amount:
          type: number
        currency:
          type: string
        description:
          type: string
        counterparty:
          type: string
        status:
          type: string
        createdAt:
          type: string
      required:
        - id
        - type
        - amount
        - currency
        - description
        - status
        - createdAt
    AdminTransactionsResponse:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/AdminTransaction"
        total:
          type: integer
      required:
        - transactions
        - total